fix(store): declare temperature as DOUBLE PRECISION and clamp negative durations
F1 (PG release blocker): llm_providers.temperature was REAL, which PG
parses as float4, while the store binds/decodes f64 — read-back failed
with 'mismatched types: f64 is not compatible with SQL type REAL'.
The failure chain was silent: ui-state import succeeded (file renamed
to .migrated), but the restart DB replay died on decrypting/decoding,
leaving GET /config empty while /health stayed green. SQLite's REAL is
8 bytes, which is why unit tests and the SQLite E2E never saw it.
0.10.0 is unreleased with no databases to migrate, so the column simply
becomes DOUBLE PRECISION (float8); SQLite maps that to the same 8-byte
REAL affinity. Decode stays f64. A new ignored PG test
(llm_providers_temperature_round_trip_on_postgres) gates the exact
read path, and cleans up its enc: rows so shared scratch databases are
not polluted with foreign-key ciphertext. migrations/0001_init.sql and
design/persistence.md document why REAL is banned for float columns.
No other float columns exist in the schema (audited: temperature is the
only one).

F2: TaskEntry::duration_ms wrapped to ~2^64 when a hand-seeded row has
created_at later than completed_at (negative i64 cast to u64). All four
millisecond-span projections (duration_ms, elapsed_ms, and the two
inline SSE elapsed calculations) now go through one millis_between
helper: saturating_sub on timestamps, clamped at 0. Unit test covers
inverted spans and clock skew.

Verified on postgres:16-alpine: both ignored store tests green; server
boot imports ui-state.toml, restart replays from the DB, and GET /config
is byte-identical across the restart (gitPlatforms/gitlab/llm intact,
temperature 0.3); an inverted-timestamp review row reports durationMs 0.

+++ migrations/0001_init.sql
@@ -93,7 +93,11 @@ CREATE TABLE llm_providers (
     api_base     TEXT NOT NULL DEFAULT '',
     api_key      TEXT NOT NULL DEFAULT '',     -- enc: 加密（新增：0.9 明文落盘）
     max_tokens   INTEGER NOT NULL DEFAULT 4096,
-    temperature  REAL NOT NULL DEFAULT 0.7,
+    -- 浮点列：必须 DOUBLE PRECISION（float8），不能用 REAL——PG 的 REAL 是
+    -- float4，而 store 层按 f64 绑定/解码（Any 驱动类型精确匹配，无隐式
+    -- 解码转换），REAL 列读回直接 mismatched types（岑静 PG E2E 实测）。
+    -- SQLite 端 DOUBLE PRECISION 同样落 REAL affinity（8 字节），两端兼容。
+    temperature  DOUBLE PRECISION NOT NULL DEFAULT 0.7,
     raw          TEXT NOT NULL DEFAULT '{}',   -- 扩展兜底：disable_thinking 等
     updated_at   TEXT NOT NULL
 );
